Move SingletonBehaviour to Exanite.Core
